[refactor] 폴더 구조 변경
- label 관련된 컴포넌트들을 labels 하위로 이동
- path alias 정리
