+++ jsconfig.json
@@ -4,6 +4,11 @@
     "paths": {
       "@src/*": ["src/*"],
       "@components/*": ["src/components/*"],
+      "@header/*": ["src/components/Header/*"],
+      "@menu/*": ["src/components/Menu/*"],
+      "@dialog/*": ["src/components/Dialog/*"],
+      "@labels/*": ["src/components/Labels/*"],
+      "@milestones/*": ["src/components/Milestones/*"],
       "@service/*": ["src/service/*"],
       "@utils/*": ["src/utils/*"],
       "@reducer/*": ["src/reducer"]
