+++ src/components/Labels/AddLabelForm.jsx
@@ -1,8 +1,8 @@
 import React, { useContext, useState } from "react";
 import styled from "styled-components";
+
 import { LabelsContext, createLabel } from "@reducer/labelReducer";
 import labelFetcher from "@service/LabelFetch";
-
 import { getRandomColor } from "@utils/utils";
 
 const lableInitial = {
