+++ src/components/Labels/Labels.jsx
@@ -1,38 +1,15 @@
-import React, { useContext, useEffect } from "react";
-import styled from "styled-components";
+import React from "react";
 
-import LabelFetch from "@service/LabelFetch.js";
-import { LabelsContext, setLabels } from "@reducer/labelReducer";
-import LabelItem from "./LabelItem";
-import LabelHead from "./LabelHead";
+import AddLabelForm from "@labels/AddLabelForm";
+import LabelList from "@labels/LabelList";
 
-const Labels = () => {
-  const { labels, labelsDispatch } = useContext(LabelsContext);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const labelList = await LabelFetch.read();
-      labelsDispatch(setLabels(labelList));
-    };
-
-    fetchData();
-  }, [labelsDispatch]);
-
-  const labelList = labels.map((label) => (
-    <LabelItem key={label.id} labelData={label} />
-  ));
-
+const Labels = ({ isNewForm, hiddenLabelForm }) => {
   return (
-    <LabelsWrapper>
-      <LabelHead labels={labels} />
-      <ul>{labelList}</ul>
-    </LabelsWrapper>
+    <>
+      <AddLabelForm visible={isNewForm} hiddenLabelForm={hiddenLabelForm} />
+      <LabelList />
+    </>
   );
 };
-
-const LabelsWrapper = styled.div`
-  margin-top: 1em;
-  border: 2px solid #f1f3f5;
-`;
 
 export default Labels;
